MSS-595: Recreate clients on each invocation as they are typically half an hour apart; so clients will always be stale
